Better handling of type aliases
This does still not solve all isuues in FoCaLiZe generated problems
because all type constants are not declared so we have to guess that
some constants are of type cc.uT (at least, this is done by Zenon
bidirectional type checker).

It becomes clear that FoCaLiZe should produce cleaner problems or
develop a tool to handle the complexity.