support more cases in THF parser
support (~)@p and (|)@p@q

see email from Geoff on 2018-2-18:
```
One of the unnatural parts of the current THF syntax is that negated
formulae have to ()ed, even if they are trivial, e.g., you have to
write ...
    fof(an,axiom,~ (p)).
... and cannot write ...
    fof(an,axiom,~ p).
The reason is that ~ is also a term that can be applied ...
    fof(an,axiom,~ @ p).
The required ()s indicate that it's being used as an operator and not as a
term.

I have had a discussion with Alex Steen (because Leo-III parses everything :-)
and devised a change that seems to be better ... require the ()s whenever a
connective is being used as a term. Alex tells me ...

> I like that solution; it is analogous to programming languages that
> require special syntax for using (infix) connectives as terms, e.g.
> is Haskell you write (||) a b for the prefix version of a || b. The
> case should not happen too often, so using parentheses is this case
> seems adequate.

Thus the examples above would become ...
    fof(an,axiom,~ p).
    fof(an,axiom,(~) @ p).
```